Fix member_of to private memory
Private memory is allocated directly in target(). This is problematic
because target_data_begin() cannot use the allocation when processing
member_of entries. Postpone their handling until the private memory
has been allocated in target().

DO NOT UPSTREAM! This change is more of a hack to make the implicit
mapping of lambdas work. The correct fix is to do all processing
in target_data_begin() and only populate the arguments in target().
This however requires a data structure for each mapping to hold
information about the just allocated data.